refactor(plugin-rsc): move Router to separate file
Extract Router and BackForwardCache classes to src/framework/router.ts
for better code organization and reusability.

- Created router.ts with Router and BackForwardCache classes
- Exported NavigationState type
- Updated entry.browser.tsx to import from router module
- entry.browser.tsx now focuses on React integration
